requirements.txt
mysql-connector-python~=8.0.28
spotipy~=2.19.0
requests~=2.27.1
aiohttp~=3.7.4.post0
beautifulsoup4~=4.10.0
discord.py~=1.7.3
yt-dlp~=2022.3.8.1